// media-futuro-pt-main/src/types/curso.ts
export interface Curso {
  id: string;
  nome_instituicao: string | null;
  tipo_ensino: string | null;
  natureza: string | null;
  nome_curso: string | null;
  grau: string | null;
  vagas_estimadas: number | null;
  media_2024: number | null;
  media_2023: number | null;
  media_2022: number | null;
}